MoonTech/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==5.2.7
django-filter==25.2
django-rest-swagger==2.2.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysqlclient==2.2.7
openapi-codec==1.3.2
packaging==25.0
pillow==11.3.0
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
simplejson==3.20.2
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
